requirements.txt
asgiref==3.3.4
Django==3.2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
mixer~=7.1.2
graphene~=2.1.8